backend/requirements.txt
# MailSpectre Requirements
# Python dependencies for the email validation backend

# Flask web framework
Flask==3.0.0

# Flask CORS support for frontend API calls
Flask-Cors==4.0.0

# DNS resolution library for domain and MX record checks
dnspython==2.4.2

# WSGI server for production deployment
gunicorn==21.2.0

# Environment variable management
python-dotenv==1.0.0
